refactor(ui): разделён ReaderView на focused reader views
ReaderView оставлен root container для ArticleScreenState, navigation flow и renderer output
toolbar actions, content surface, body block presentation, overscroll indicators и cached image loading вынесены в отдельные файлы
