+++ RSSReader/Views/ArticleScreen/ArticleScreenActionHandlers.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct ArticleScreenActionHandlers {
+    let toggleReadStatus: () -> Void
+    let toggleStarredStatus: () -> Void
+    let openSourceArticle: () -> Void
+    let bodyLinkTapped: (URL) -> Void
+}
